tragent_backend/requirements.txt
asgiref==3.8.0
beautifulsoup4==4.12.3
contourpy==1.2.0
cycler==0.12.1
Django==5.0.3
djangorestframework==3.15.0
fonttools==4.50.0
geojson==3.1.0
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.3
numpy==1.26.4
OSMPythonTools==0.3.5
packaging==24.0
pandas==2.2.1
pillow==10.2.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
ujson==5.9.0
xarray==2024.2.0
